restructuring code to work with Azure Functions
