--- app/__init__.py
@@ -1,5 +1,0 @@
-from quart import Quart
-
-app = Quart(__name__)
-
-from . import controllers, services, models
