+++ core/models/__init__.py
@@ -1,0 +1,1 @@
+from . import food, store, store_brand
