+++ pickup/function.json
@@ -1,11 +1,13 @@
 {
   "scriptFile": "__init__.py",
+
   "bindings": [
     {
       "authLevel": "function",
       "type": "httpTrigger",
       "direction": "in",
       "name": "req",
+      "route": "pickup/{zipcode}",
       "methods": [
         "get",
         "post"
