+++ requirements.txt
@@ -1,1 +1,7 @@
+aiohttp==3.6.2
+async-timeout==3.0.1
+asyncio==3.4.3
 azure-functions==1.2.0
+dataclasses-json==0.4.2
+jsonpickle==1.3
+requests==2.20.0
